% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascule-package.R
\docType{package}
\name{bascule-package}
\alias{bascule}
\alias{bascule-package}
\title{bascule: bascule}
\description{
This packages provides the bascule (Bayesian signature learning including a catalogue) algorithm to detect single-nucleotide substituion signatures in cancer. The main feature of bascule is to provide a full Bayesian hierarchical formulation based on non-negative matrix factorisation, leveraging a semi-supervised trainig set (i.e., a catalogue). The tool is specifically developed to detect de novo mutational signatures that are different from the pre-defined reference catalogue.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caravagnalab/bascule}
  \item \url{https://caravagnalab.github.io/bascule/}
}

}
\author{
\strong{Maintainer}: Elena Buscaroli \email{ele.buscaroli@gmail.com}

Authors:
\itemize{
  \item Giulio Caravagna \email{gcaravagn@gmail.com}
}

}
\keyword{internal}
