% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilica-package.R
\docType{package}
\name{basilica-package}
\alias{basilica}
\alias{basilica-package}
\title{basilica: basilica}
\description{
This packages provides the basilica (Bayesian signature learning including a catalogue) algorithm to detect single-nucleotide substituion signatures in cancer. The main feature of basilica is to provide a full Bayesian hierarchical formulation based on non-negative matrix factorisation, leveraging a semi-supervised trainig set (i.e., a catalogue). The tool is specifically developed to detect de novo mutational signatures that are different from the pre-defined reference catalogue.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caravagnalab/basilica}
  \item \url{https://caravagnalab.github.io/basilica/}
}

}
\author{
\strong{Maintainer}: Azad Sadr \email{azad.sadr.h@gmail.com}

Authors:
\itemize{
  \item Giulio Caravagna \email{gcaravagn@gmail.com}
}

}
\keyword{internal}
