% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_data}
\alias{breast_data}
\title{Analysis of breast tumours}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(breast_data)
}
\description{
Input SBS and DBS counts and BASCULE fits from 2682 breast tumours
released in "Substitution mutational signatures in whole-genome–sequenced
cancers in the UK population."
Degasperi et al.Science 376, Issue 6591, 2022. PMID: 35949260
}
\keyword{datasets}
