% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_QC.R
\name{convert_dn_names}
\alias{convert_dn_names}
\title{Function to map de novo signatures to input catalogues}
\usage{
convert_dn_names(x, x.simul = NULL, reference_cat = NULL, cutoff = 0.8)
}
\arguments{
\item{x}{Object of class "bascule_obj"}

\item{x.simul}{Another object of class "bascule_obj". If present, `x` de novo signatures will be mapped to `x_simul` ones}

\item{reference_cat}{List of reference catalogues. If not `NULL`, `x` de novo signatures will be mapped to these catalogues.}

\item{cutoff}{Threshold for the cosine similarity.}
}
\value{
Modified version of `x` with the mapped de novo names renamed.
}
\description{
Function to map de novo signatures to input catalogues
}
