% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_cn}
\alias{COSMIC_cn}
\title{COSMIC catalogue for CN (version 3.4, GRCh37)}
\format{
An object of class \code{data.frame} with 24 rows and 48 columns.
}
\usage{
data(COSMIC_cn)
}
\description{
COSMIC catalogue for CN (version 3.4, GRCh37)
}
\keyword{datasets}
