% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_filt_merged}
\alias{COSMIC_filt_merged}
\title{COSMIC catalogue (version XXXX)}
\format{
add
}
\usage{
data(COSMIC_filt_merged)
}
\description{
Filtered version of the COSMIC catalogue. Only validated signatures are present.
For each signature (besides SBS5) the contexts with probability mass below 0.02 are set to 0
and the probabilities are re-normalized over contexts.
Moreover, all signatures with high cosine similarity are merged.
}
\keyword{datasets}
