% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_indels}
\alias{COSMIC_indels}
\title{COSMIC catalogue for Indels (version 3.4, GRCh37)}
\format{
An object of class \code{data.frame} with 18 rows and 83 columns.
}
\usage{
data(COSMIC_indels)
}
\description{
COSMIC catalogue for Indels (version 3.4, GRCh37)
}
\keyword{datasets}
