% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_sbs}
\alias{COSMIC_sbs}
\title{COSMIC catalogue for SBS (version 3.4, GRCh37)}
\format{
add
}
\usage{
data(COSMIC_sbs)
}
\description{
COSMIC catalogue for SBS (version 3.4, GRCh37)
}
\keyword{datasets}
