% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_sbs_filt}
\alias{COSMIC_sbs_filt}
\title{COSMIC catalogue for SBS filtered (version 3.4, GRCh37)}
\format{
add
}
\usage{
data(COSMIC_sbs_filt)
}
\description{
Filtered version of the COSMIC catalogue. Only validated signatures are present.
For each signature (besides SBS5) the contexts with probability mass below 0.02 are set to 0
and the probabilities are re-normalized over contexts.
}
\keyword{datasets}
