% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Degasperi_catalogue}
\alias{Degasperi_catalogue}
\title{Degasperi catalogue (version XXXX)}
\format{
....
}
\usage{
data(Degasperi_catalogue)
}
\description{
D......
}
\examples{
data(Degasperi_catalogue)
print(Degasperi_catalogue)
}
\keyword{datasets}
