% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Degasperi_SBS}
\alias{Degasperi_SBS}
\title{Degasperi SBS catalogue}
\format{
An object of class \code{data.frame} with 120 rows and 96 columns.
}
\usage{
data(Degasperi_SBS)
}
\description{
SBS signatures extracted and released in "Substitution mutational signatures
in whole-genome–sequenced cancers in the UK population."
Degasperi et al.Science 376, Issue 6591, 2022. PMID: 35949260
}
\keyword{datasets}
