% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Degasperi_SBS_GEL_PCAWG_HW}
\alias{Degasperi_SBS_GEL_PCAWG_HW}
\title{SBS data for PMID: 35949260}
\format{
An object of class \code{list} of length 21.
}
\usage{
data(Degasperi_SBS_GEL_PCAWG_HW)
}
\description{
SBS data released with the paper

Substitution mutational signatures in whole-genome–sequenced cancers in the UK population.
Degasperi et al.Science 376, Issue 6591, 2022. PMID: 35949260
}
\examples{
data(Degasperi_SBS_GEL_PCAWG_HW)
print(Degasperi_SBS_GEL_PCAWG_HW)
}
\keyword{datasets}
