% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_input}
\alias{example_input}
\title{Example input SBS data (synthetic)}
\format{
An object of class \code{data.frame} with 5 rows and 96 columns.
}
\usage{
data(example_input)
}
\description{
Description
}
\keyword{datasets}
