% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate.cohort}
\alias{generate.cohort}
\title{Title}
\usage{
generate.cohort(
  reference_path,
  ratio,
  targetX,
  inputX,
  similarity_limit,
  groups,
  mut_range,
  seed = NULL,
  num_data
)
}
\arguments{
\item{reference_path}{e.g., '/home/file.csv', path for reference catalogue file in csv format}

\item{ratio}{e.g., 0.7, ratio between reference and denovo catalogues, both derived from splitting main reference catalogue in two parts}

\item{targetX}{e.g., c(a,b), a: No. of fixed signatures and b: No. of denovo signatures in data}

\item{inputX}{e.g., c(a,b), a: No. of fixed signatures and b: No. of denovo signatures in input}

\item{similarity_limit}{similarity between signatures in catalogues (trying to include non-similar signatures)}

\item{groups}{e.g., rep(1, n), n is number of patients in data (1st one is always one)}

\item{mut_range}{# range of mutations counts in patients}

\item{seed}{obvious}

\item{num_data}{e.g., 50, No. of the data (rows) in cohort. each data is a mutational catalogues comprising several patients}
}
\value{
a cohort includes several data
}
\description{
Title
}
