% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_catalogue_signatures}
\alias{get_catalogue_signatures}
\title{get catalogue signatures}
\usage{
get_catalogue_signatures(x)
}
\arguments{
\item{x}{basilica object}
}
\value{
inferred signatures which are from reference catalogue
}
\description{
get catalogue signatures
}
