% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_denovo_signames}
\alias{get_denovo_signames}
\title{Get denovo signatures names}
\usage{
get_denovo_signames(x, types = get_types(x))
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to retrieve signames for.}
}
\value{
list of de novo signature names with names equal to `types`.
}
\description{
Get denovo signatures names
}
