% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_denovo_signatures}
\alias{get_denovo_signatures}
\title{get de novo signatures}
\usage{
get_denovo_signatures(x)
}
\arguments{
\item{x}{basilica object}
}
\value{
inferred signatures which are not included in reference catalogue
}
\description{
get de novo signatures
}
