% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_exposure}
\alias{get_exposure}
\title{get exposure matrix}
\usage{
get_exposure(x, long = FALSE)
}
\arguments{
\item{x}{basilica object}

\item{long}{if TRUE return the long format exposure matrix (default=FALSE)}
}
\value{
a data.frame where rows are samples and columns are inferred signature profiles
}
\description{
get exposure matrix
}
