% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_fixed_signames}
\alias{get_fixed_signames}
\title{Get reference signatures names}
\usage{
get_fixed_signames(x, types = get_types(x))
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to retrieve signames for.}
}
\value{
list of reference signature names with names equal to `types`.
}
\description{
Get reference signatures names
}
