% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_input}
\alias{get_input}
\title{Get input data}
\usage{
get_input(
  x,
  types = get_types(x),
  samples = get_samples(x),
  clusters = get_cluster_labels(x),
  matrix = FALSE,
  reconstructed = FALSE,
  add_groups = FALSE,
  by_sigs = FALSE
)
}
\arguments{
\item{x}{basilica object.}

\item{types}{List of variant types to subset}

\item{samples}{List of samples to subset}

\item{clusters}{List of clusters to subset}

\item{matrix}{Logical. If `TRUE`, the output will be in wide format.}

\item{reconstructed}{Logical. If `TRUE`, the reconstructed counts, i.e., computed as alpha x beta, will be returned.}

\item{add_groups}{Logical. If `TRUE` and clustering has been performed, each sample will be annotated with the assigned group.}

\item{by_sigs}{Logical. If `TRUE`, the number of mutations generated by each signature will be returned.}
}
\value{
list of long or wide-formatted mutations counts.
}
\description{
Get input data
}
\examples{
get_input(example_fit)
}
