% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_QC.R
\name{get_n_denovo}
\alias{get_n_denovo}
\title{Get number of de novo signatures}
\usage{
get_n_denovo(x)
}
\arguments{
\item{x}{bascule object}
}
\value{
list with an integer number of de novo signatures for each signature type
}
\description{
Get number of de novo signatures
}
