% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_signames}
\alias{get_signames}
\title{Get signatures names}
\usage{
get_signames(x, types = get_types(x))
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to retrieve signames for.}
}
\value{
list of signature names with names equal to `types`.
}
\description{
Get signatures names
}
