% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_signatures}
\alias{get_signatures}
\title{Get signatures}
\usage{
get_signatures(x, types = get_types(x), matrix = FALSE)
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to retrieve signames for.}

\item{matrix}{Logical. If `TRUE`, it will return the signatures in wide format.}
}
\value{
Signature matrix in long or wide format names with names equal to `types`.
}
\description{
Get signatures
}
