% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environment.R
\name{have_loaded_env}
\alias{have_loaded_env}
\title{Check if there is a loaded \code{conda} environment.}
\usage{
have_loaded_env()
}
\value{
Boolean, \code{TRUE} if an environment is loaded, \code{FALSE} otherwise.
}
\description{
Function to check if a \code{conda} environment is already loaded.
}
