% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environment.R
\name{have_python_deps}
\alias{have_python_deps}
\title{Check if \code{Python} packages are installed in the environment.}
\usage{
have_python_deps(envname = "", py_pkgs = c("pybasilica"))
}
\arguments{
\item{envname}{the name of the environment to check. If empty, the function will check the
currently loaded environment.}

\item{py_pkgs}{a list or vector of \code{Python} packages.}
}
\value{
a list of Boolean. For each input package, \code{TRUE} if the package is
installed, \code{FALSE} otherwise.
}
\description{
Function to check if one or more \code{Python} packages are
present in a \code{conda} environment.
}
