% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clustering.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Function to merge similar clusters}
\usage{
merge_clusters(x, cutoff = 0.8)
}
\arguments{
\item{x}{basilica object.}

\item{cutoff}{Minimum value of similarity to merge two clusters.}
}
\value{
basilica object.
}
\description{
This function will iteratively merge clusters that result similar in the centroid.
The merging will stop as soon as the cosine similarity between all pairs of clusters
is below `cutoff`
}
