% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_S3.R
\name{plot.bascule_obj}
\alias{plot.bascule_obj}
\title{Plot for class \code{'bascule_obj'}.}
\usage{
\method{plot}{bascule_obj}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'bascule_obj'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
The default plot is the CNA segments in wide format
}
\examples{
\dontrun{
}
}
