% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposures.R
\name{plot_centroids}
\alias{plot_centroids}
\title{Plot clustering centroids}
\usage{
plot_centroids(
  x,
  types = get_types(x),
  clusters = get_cluster_labels(x),
  cls = NULL,
  sort_by = NULL,
  exposure_thr = 0,
  quantile_thr = 0,
  signatures_list = get_signames(x),
  ...
)
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to visualize.}

\item{clusters}{List of clusters to visualize.}

\item{cls}{Custom color palette for signatures.}

\item{sort_by}{Signature to sort patients' exposures by.}

\item{exposure_thr}{Only signatures with exposures greater than `exposure_thr` in all samples will be highlighted.}

\item{quantile_thr}{add}

\item{signatures_list}{add}

\item{...}{Additional arguments}
}
\value{
ggplot object.
}
\description{
Plot clustering centroids
}
