% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_fits.R
\name{plot_compare_fits}
\alias{plot_compare_fits}
\title{Title}
\usage{
plot_compare_fits(x, y, similarity_cutoff = 0.4)
}
\arguments{
\item{x}{add}

\item{y}{add}

\item{similarity_cutoff}{add}
}
\value{
add
}
\description{
Title
}
