% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Function to visualize the samples mutation profiles.}
\usage{
plot_data(x, sample_ids = NULL, what = "SBS", context = T)
}
\arguments{
\item{x}{Basilica object.}

\item{sample_ids}{Array with the samples IDs to show.}

\item{what}{Character among \code{["SBS","DBS","ID","CNV"]} reporting the type of signature.}

\item{context}{Logical. If set to \code{TRUE}, the contexts on the x axis will be shown.}
}
\value{
plot
}
\description{
creates bar plot of data profile for each sample,
where x-axis are 96 substitution bases and y-axis are their relative contribution.
}
