% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_exposure}
\alias{plot_exposure}
\title{plot relative exposure matrix barplot}
\usage{
plot_exposure(x)
}
\arguments{
\item{x}{basilica object}
}
\value{

}
\description{
plot relative exposure matrix barplot
}
