% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposures.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Function to visualize the estimated exposures}
\usage{
plot_exposures(
  x,
  types = get_types(x),
  samples = get_samples(x),
  clusters = get_cluster_labels(x),
  sample_name = F,
  sigs_levels = NULL,
  cls = NULL,
  add_centroid = FALSE,
  muts = FALSE,
  sort_by = NULL,
  elim_thr = 0
)
}
\arguments{
\item{x}{basilica object.}

\item{types}{List of variant types to visualize.}

\item{samples}{List of samples to visualize.}

\item{clusters}{List of clusters to visualize.}

\item{sample_name}{Logical. If `TRUE`, sample names will be reported on the x axis.}

\item{sigs_levels}{Deprecated.}

\item{cls}{Custom color palette for signatures.}

\item{add_centroid}{Logical. If `TRUE`, also clustering's centroids will be plotted.}

\item{muts}{Deprecated.}

\item{sort_by}{Signature to sort patients' exposures by.}

\item{elim}{merge all signatures with low exposure as single signature called 'lowexp'}
}
\value{
ggplot object.
}
\description{
Function to visualize the estimated exposures
}
\examples{
plot_exposures(example_fit)
}
