% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC.R
\name{plot_gradient_norms}
\alias{plot_gradient_norms}
\title{Plot parameters gradients norms}
\usage{
plot_gradient_norms(x, types = get_types(x))
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of types to visualise gradient norms for.}
}
\value{
patchwork object
}
\description{
Visualise the norms of the parameters gradients computed during the inference.
}
