% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_mutations}
\alias{plot_mutations}
\title{Function to plot the overall mutation profile}
\usage{
plot_mutations(x, sampleIDs = NULL, by_sig = FALSE, reconstructed = FALSE)
}
\arguments{
\item{x}{Basilica object.}

\item{sampleIDs}{List containing the sample IDs to consider in the counts.
If set to \code{NULL}, all samples will be considered.}

\item{by_sig}{Logical, if set to \code{TRUE} each context will report
its signatures contribution.}

\item{reconstructed}{whether to plot the reconstructed or input matrix}
}
\value{
ggplot object
}
\description{
Function to plot the overall mutation profile
}
