% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC.R
\name{plot_posterior_probs}
\alias{plot_posterior_probs}
\title{Plot posterior probabilities}
\usage{
plot_posterior_probs(x, samples = get_samples(x))
}
\arguments{
\item{x}{bascule object}

\item{samples}{List of samples to visualise.}
}
\value{
ggplot object
}
\description{
Plot the posterior probabilities over the assignments.
}
