% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_QC.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot model selection scores}
\usage{
plot_scores(x, types = get_types(x), remove_outliers = FALSE)
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to visualize.}

\item{remove_outliers}{Logical. If `TRUE`, outliers in each score will be removed.}
}
\value{
ggplot2 object.
}
\description{
Function to plot the model selection scores for NMF on each variant type and clustering.
Reported scores are BIC and negative log-likelihood.
}
