% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similarity_reference.R
\name{plot_similarity_reference}
\alias{plot_similarity_reference}
\title{Title}
\usage{
plot_similarity_reference(
  x,
  reference = NULL,
  similarity_cutoff = 0.4,
  by_subs = FALSE
)
}
\arguments{
\item{similarity_cutoff}{}
}
\description{
Title
}
