% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_S3.R
\name{print.basilica_obj}
\alias{print.basilica_obj}
\title{Print for class \code{'basilica_obj'}.}
\usage{
\method{print}{basilica_obj}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'basilica_obj'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'basilica_obj'}.
}
