% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_refine.R
\name{refine_denovo_signatures}
\alias{refine_denovo_signatures}
\title{De novo signatures refinement.}
\usage{
refine_denovo_signatures(x, types = get_types(x))
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to perform de novo refinement on.}
}
\value{
bascule object.
}
\description{
Function to refine the inferred de novo signatures.
The function performs a linear combination between de novo and reference
signatures. If a de novo signature can be explained as a linear combination
of one or more reference signatures, it will be removed and its exposures
will be distributed among the similar signatures.
}
