% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skin_fit}
\alias{skin_fit}
\title{Skin cohort BASCULE fit}
\format{
An object of class \code{bascule_obj} of length 3.
}
\usage{
data(skin_fit)
}
\description{
BASCULE fit for skin cohort data used in the survival analysis.
}
\keyword{datasets}
