% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skin_metadata}
\alias{skin_metadata}
\title{Skin cohort metadata}
\format{
An object of class \code{data.frame} with 1832 rows and 7 columns.
}
\usage{
data(skin_metadata)
}
\description{
Clinical data for the skin cancer cohort used in the survival analysis.
}
\keyword{datasets}
