% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_data}
\alias{synthetic_data}
\title{Analysis of a synthetic cohort}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(synthetic_data)
}
\description{
Input SBS and DBS counts and BASCULE fits for a cohort of 150 samples whose
counts and exposures have been generated from the BASCULE generative model.
}
\keyword{datasets}
