% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_environment.R
\name{which_conda_env}
\alias{which_conda_env}
\title{Retrieve the name of the currently loaded environment.}
\usage{
which_conda_env()
}
\value{
a \code{character} corresponding to the name of the loaded environment.
}
\description{
Function to retrieve the name of the currently loaded \code{conda} environment.
}
