% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bascule-package.R
\docType{package}
\name{bascule-package}
\alias{bascule}
\alias{bascule-package}
\title{bascule: bascule}
\description{
This packages provides the bascule (Bayesian inference and clustering with mutational signatures and an established catalogue) algorithm to detect single-nucleotide substituion signatures in cancer. bascule provides a full Bayesian formulation based on non-negative matrix factorisation, leveraging a catalogue of known signatures. The tool is specifically developed to detect de novo mutational signatures of any type (SBS, DBS, IDS, ...) that are different from the pre-defined reference catalogue and to perform clustering of samples from all the signature types jointly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caravagnalab/bascule}
  \item \url{https://caravagnalab.github.io/bascule/}
}

}
\author{
\strong{Maintainer}: Elena Buscaroli \email{ele.buscaroli@gmail.com}

Authors:
\itemize{
  \item Giulio Caravagna \email{gcaravagn@gmail.com}
}

}
\keyword{internal}
