% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_catalogue}
\alias{COSMIC_catalogue}
\title{COSMIC catalogue (version XXXX)}
\format{
add
}
\usage{
data(COSMIC_catalogue)
}
\description{
COSMIC catalogue (version XXXX)
}
\keyword{datasets}
