% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COSMIC_dbs}
\alias{COSMIC_dbs}
\title{COSMIC catalogue for DBS (version 3.4, GRCh37)}
\format{
An object of class \code{data.frame} with 20 rows and 78 columns.
}
\usage{
data(COSMIC_dbs)
}
\description{
COSMIC catalogue for DBS (version 3.4, GRCh37)
}
\keyword{datasets}
