% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilica.R
\name{fit}
\alias{fit}
\title{fit basilica model.}
\usage{
fit(
  x,
  k,
  py = NULL,
  reference_catalogue = basilica::COSMIC_catalogue,
  input_catalogue = basilica::COSMIC_catalogue["SBS1", ],
  filtered_cat = FALSE,
  cohort = "MyCohort",
  lr = 0.01,
  steps = 500,
  max_iterations = 20,
  blacklist = NULL,
  phi = 0.05,
  delta = 0.9,
  filt_pi = 0.1,
  groups = NULL,
  lambda_rate = NULL,
  sigma = FALSE,
  CUDA = FALSE,
  compile = FALSE,
  enforce_sparsity = FALSE,
  store_parameters = FALSE,
  regularizer = "cosine",
  reg_weight = 1,
  reg_bic = FALSE,
  cosine_by_subs = FALSE
)
}
\arguments{
\item{x}{input mutational counts data (data.frame; rows as samples and columns as 96 mutational categories)}

\item{k}{vector of possible number of de novo signatures to infer}

\item{reference_catalogue}{a catalog of reference signatures that basilica will use to compare input and de novo signatures (COSMIC catalogue by default)}

\item{input_catalogue}{input signature profiles, NULL by default}

\item{cohort}{add}

\item{lr}{stochastic variational inference learning rate}

\item{steps}{number of gradient steps}

\item{max_iterations}{add}

\item{blacklist}{add}

\item{phi}{threshold to discard the signature based on its value in exposure matrix}

\item{delta}{threshold to consider inferred signature as COSMIC signature}

\item{filt_pi}{threshold for COSMIC signature weight to be included as basis of the denovo signatures}

\item{groups}{vector of discrete labels with one entry per sample, it defines the groups that will be considered by basilica}

\item{lambda_rate}{add}

\item{sigma}{add}

\item{enforce_sparsity}{use Laplace prior over exposure weights (bayesian LASSO)}
}
\value{
inferred exposure matrix, inferred signatures from reference catalogue and inferred de novo (not from reference catalogue) signatures
}
\description{
fit the model and infer the underlying signatures and their contributions in the mutational catalogue counts.
}
