% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFinder.R
\name{fit}
\alias{fit}
\title{Title}
\usage{
fit(
  x,
  input_catalog = NULL,
  k = 0:5,
  reference_catalog = basilica::COSMIC,
  lr = 0.05,
  steps_per_iter = 500,
  fixedLimit = 0.05,
  denovoLimit = 0.9
)
}
\arguments{
\item{x}{input mutational counts data (data.frame; rows as samples and columns as 96 mutational categories)}

\item{input_catalog}{input signature profiles, NULL by default}

\item{k}{vector of possible number of de novo signatures to infer}

\item{reference_catalog}{a catalog of reference signatures that basilica will use to compare input and de novo signatures}

\item{fixedLimit}{threshold to discard the signature based on its value in exposure matrix}

\item{denovoLimit}{threshold to consider inferred signature as COSMIC signature}
}
\value{
inferred exposure matrix, inferred COSMIC signatures and inferred de novo (not from referencecatalog) signatures
}
\description{
Title
}
