% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureFinder.R
\name{fit}
\alias{fit}
\title{fit basilica model.}
\usage{
fit(
  x,
  groups = NULL,
  input_catalogue = NULL,
  reference_catalogue = basilica::COSMIC_catalogue,
  k = 0:5,
  lr = 0.05,
  steps = 500,
  phi = 0.05,
  delta = 0.9
)
}
\arguments{
\item{x}{input mutational counts data (data.frame; rows as samples and columns as 96 mutational categories)}

\item{groups}{vector of discrete labels with one entry per sample, it defines the groups that will be considered by basilica}

\item{k}{vector of possible number of de novo signatures to infer}

\item{lr}{stochastic variational inference learning rate}

\item{steps}{number of gradient steps}

\item{phi}{threshold to discard the signature based on its value in exposure matrix}

\item{delta}{threshold to consider inferred signature as COSMIC signature}

\item{input_catalog}{input signature profiles, NULL by default}

\item{reference_catalog}{a catalog of reference signatures that basilica will use to compare input and de novo signatures (COSMIC catalogue by default)}
}
\value{
inferred exposure matrix, inferred signatures from reference catalogue and inferred de novo (not from reference catalogue) signatures
}
\description{
fit the model and infer the underlying signatures and their contributions in the mutational catalogue counts.
}
