% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_clustering}
\alias{fit_clustering}
\title{Fit clustering}
\usage{
fit_clustering(
  x,
  cluster,
  hyperparameters = NULL,
  lr = 0.005,
  optim_gamma = 0.1,
  n_steps = 3000,
  py = NULL,
  enumer = "parallel",
  nonparametric = TRUE,
  autoguide = TRUE,
  CUDA = TRUE,
  compile = FALSE,
  store_parameters = FALSE,
  store_fits = TRUE,
  seed_list = c(10)
)
}
\arguments{
\item{x}{Bascule object with signatures deconvolution performed.}

\item{cluster}{Maximum number of clusters.}

\item{hyperparameters}{List of hyperparameters passed to the NMF and clustering models.}

\item{lr}{Learning rate for SVI optimizer.}

\item{optim_gamma}{Deprecated.}

\item{n_steps}{Number of steps for the inference.}

\item{py}{User-installed version of \code{pybascule} package}

\item{enumer}{Enumeration used for clustering (either `parallel` or `sequential`).}

\item{nonparametric}{Deprecated. The model only works in nonparametric way.}

\item{autoguide}{Logical. If `TRUE`, the clustering model will use the Pyro autoguide.}

\item{CUDA}{Logical. If `TRUE` and a GPU is available, the models will run on GPU.}

\item{compile}{Deprecated.}

\item{store_parameters}{Logical. If `TRUE`, parameters at every step of inference will be stored in the object.}

\item{store_fits}{Logical. If `TRUE`, all tested fits, i.e., for every value of `K`, will be stored in the object.}

\item{seed_list}{List of seeds used for every input configuration.}
}
\value{
Bascule object.
}
\description{
Fit clustering
}
