% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_catalogue_signatures}
\alias{get_catalogue_signatures}
\title{get catalogue signatures}
\usage{
get_catalogue_signatures(x, long = FALSE)
}
\arguments{
\item{x}{basilica object}
}
\value{
a data.frame where rows are inferred signatures (included in reference catalogue) and columns are 96 substitution bases.
}
\description{
get catalogue signatures
}
