% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_denovo_signatures}
\alias{get_denovo_signatures}
\title{get de novo signatures}
\usage{
get_denovo_signatures(x, long = FALSE)
}
\arguments{
\item{x}{basilica object}
}
\value{
a data.frame where rows are inferred signatures (not included in reference catalogue) and columns are 96 substitution bases.
}
\description{
get de novo signatures
}
