% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_exposure}
\alias{get_exposure}
\title{Get exposures}
\usage{
get_exposure(
  x,
  types = get_types(x),
  samples = get_samples(x),
  clusters = get_cluster_labels(x),
  add_groups = FALSE,
  matrix = FALSE
)
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to retrieve signames for.}

\item{samples}{List of samples to report exposures for.}

\item{clusters}{List of cluster labels to report exposures for.}

\item{add_groups}{Logical. If `TRUE` it will add a column with the sample's group label.}

\item{matrix}{Logical. If `TRUE`, it will return the signatures in wide format.}
}
\value{
Exposures matrix in long or wide format names with names equal to `types`.
}
\description{
Get exposures
}
