% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutations.R
\name{plot_data}
\alias{plot_data}
\title{Function to visualize the input and reconstructed mutation counts.}
\usage{
plot_data(
  x,
  types = get_types(x),
  samples = get_samples(x),
  clusters = get_cluster_labels(x),
  reconstructed = TRUE,
  color_by_sigs = FALSE,
  plot_by_sample = FALSE
)
}
\arguments{
\item{x}{bascule object.}

\item{types}{List of variant types to visualize.}

\item{samples}{List of samples to visualize.}

\item{clusters}{List of clusters to visualize.}

\item{reconstructed}{Logical. If `TRUE`, the reconstructed counts, i.e., computed as alpha x beta, will be visualized.}

\item{color_by_sigs}{Logical. If `TRUE`, each context will report the number of mutations split by signature.}

\item{plot_by_sample}{Logical. If `TRUE`, the function will return a list of ggplot objects with the counts for each sample.}
}
\value{
one or a list of ggplot objects.
}
\description{
Function to visualize the input and reconstructed mutation counts.
}
