% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_exposure}
\alias{plot_exposure}
\title{plot exposure matrix}
\usage{
plot_exposure(x, sort_by = NULL)
}
\arguments{
\item{x}{basilica object}
}
\value{
plot
}
\description{
creates bar plot of relative exposure matrix, where x-axis are samples and y-axis are their relative contribution.
}
