% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Plot exposure matrix}
\usage{
plot_exposures(
  x,
  sample_name = F,
  sigs_levels = NULL,
  cls = NULL,
  flip_coord = F,
  muts = FALSE,
  sampleIDs = NULL,
  sort_by = NULL
)
}
\arguments{
\item{x}{Basilica object}

\item{sample_name}{Logical. If set to \code{TRUE}, the sample IDs will be
reported on the x axis.}

\item{cls}{add}

\item{flip_coord}{add}

\item{levels}{add}
}
\value{
plot
}
\description{
Creates bar plot of relative exposure matrix, where x-axis are
             samples and y-axis are their relative contribution.
}
