% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Title}
\usage{
plot_signatures(
  x,
  useRowNames = FALSE,
  xlabels = FALSE,
  denovoSignature = TRUE
)
}
\arguments{
\item{x}{basilica object}

\item{useRowNames}{using signature names from data.frame}

\item{xlabels}{axis label}

\item{denovoSignature}{plotting inferred signaturese from de-novo or catalog}
}
\value{

}
\description{
Title
}
