% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_signatures}
\alias{plot_signatures}
\title{plot signatures}
\usage{
plot_signatures(x, denovo = TRUE)
}
\arguments{
\item{x}{basilica object}

\item{denovoSignature}{if TRUE, plots inferred de-novo signatures, otherwise plots inferred catalogue signatures}
}
\value{
plot
}
\description{
creates bar plot of inferred signature profiles, where x-axis are 96 substitution bases and y-axis are their relative contribution.
}
