% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_signatures}
\alias{plot_signatures}
\title{plot signatures}
\usage{
plot_signatures(x, Type = c("De novo", "Catalogue"), highlight = 0.05)
}
\arguments{
\item{x}{basilica object}
}
\value{
plot
}
\description{
creates bar plot of inferred signature profiles, where x-axis are 96 substitution bases and y-axis are their relative contribution.
}
