% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures.R
\name{plot_signatures}
\alias{plot_signatures}
\title{Function to visualize the estimated and reference signatures.}
\usage{
plot_signatures(
  x,
  types = get_types(x),
  context = T,
  cls = NULL,
  signames = get_signames(x)
)
}
\arguments{
\item{x}{basilica object.}

\item{types}{List of variant types to visualize.}

\item{context}{Logical. If `TRUE`, context names will be reported on the x axis.}

\item{cls}{Custom color palette for signatures.}

\item{signames}{List of signatures to visualize.}
}
\value{
ggplot object.
}
\description{
Function to visualize the estimated and reference signatures.
}
\examples{
plot_signatures(example_fit)
}
